+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,0 +1,23 @@
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Cards;
+
+public class Player extends Participant {
+
+    public Player(Name name, Cards cards) {
+        super(name, cards);
+    }
+
+    @Override
+    public boolean isFinished() {
+        return cards.isBust() || cards.isBlackjack();
+    }
+
+    @Override
+    public String toString() {
+        return "Player{" +
+                "name=" + name +
+                ", cards=" + cards +
+                '}';
+    }
+}
