--- src/main/java/domain/participant/Role.java
@@ -1,4 +1,0 @@
-package domain.participant;
-
-public interface Role {
-}
